Refactor analytics logic into modular clients
Split analytics implementation into separate modules for PostHog, GA4, REO, and cookie consent management. Centralized shared types and utilities, improving maintainability and separation of concerns. The main analytics utility now delegates to these clients, simplifying configuration and event tracking.
